requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
Flask>=2.3.0
Flask-Session>=0.5.0
Pillow>=10.0.0
aiohttp>=3.8.0
requests>=2.31.0